linter, solidity prettier

+++ contracts/Migrations.sol
@@ -1,20 +1,18 @@
 pragma solidity ^0.5.11;
 
-
 contract Migrations {
     address public owner;
-    uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
     modifier restricted() {
-        if (msg.sender == owner)
-            _;
+        if (msg.sender == owner) _;
     }
 
     constructor() public {
         owner = msg.sender;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         last_completed_migration = completed;
     }
